+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Highlights
 
 ## Changes
+- Set the use of system fonts in the configuration file and not only as a CLI flag
 
 ## Fixes